rename some source files to make more readable
