+++ async_simple/coro/SyncAwait.h
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifndef ASYNC_SIMPLE_CORO_LAZY_HELPER_H
-#define ASYNC_SIMPLE_CORO_LAZY_HELPER_H
+#ifndef ASYNC_SIMPLE_CORO_SYNC_AWAIT_H
+#define ASYNC_SIMPLE_CORO_SYNC_AWAIT_H
 
 #include <async_simple/Common.h>
 #include <async_simple/experimental/coroutine.h>
@@ -55,7 +55,5 @@ inline auto syncAwait(LazyType &&lazy, Executor *ex) {
 
 }  // namespace coro
 }  // namespace async_simple
-
-#include <async_simple/coro/Collect.h>
 
 #endif
