+++ async_simple/coro/test/FutureAwaiterTest.cpp
@@ -15,6 +15,7 @@
  */
 #include <async_simple/coro/FutureAwaiter.h>
 #include <async_simple/coro/Lazy.h>
+#include <async_simple/coro/SyncAwait.h>
 #include <async_simple/test/unittest.h>
 #include <chrono>
 #include <thread>
