+++ async_simple/coro/test/LazyTest.cpp
@@ -22,7 +22,9 @@
 
 #include <iostream>
 
+#include <async_simple/coro/Collect.h>
 #include <async_simple/coro/Lazy.h>
+#include <async_simple/coro/SyncAwait.h>
 #include <async_simple/coro/test/Time.h>
 #include <async_simple/executors/SimpleExecutor.h>
 #include <async_simple/experimental/coroutine.h>
