+++ async_simple/coro/test/ViaCoroutineTest.cpp
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 #include <async_simple/coro/Lazy.h>
+#include <async_simple/coro/SyncAwait.h>
 #include <async_simple/executors/SimpleExecutor.h>
 #include <async_simple/test/unittest.h>
 #include <exception>
