+++ demo_example/CountChar.cpp
@@ -13,11 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include <async_simple/coro/Lazy.h>
+#include <async_simple/coro/SyncAwait.h>
 #include <algorithm>
 #include <fstream>
 #include <iostream>
 #include <string>
-#include "async_simple/coro/Lazy.h"
+#include <vector>
 
 using namespace async_simple::coro;
