+++ demo_example/asio_util.hpp
@@ -17,6 +17,7 @@
 #define ASYNC_SIMPLE_DEMO_ASIO_UTIL_H
 
 #include <async_simple/coro/Lazy.h>
+#include <async_simple/coro/SyncAwait.h>
 #include <async_simple/executors/SimpleExecutor.h>
 #include <cstdlib>
 #include <iostream>
